Implement new anomaly/interfacing systems (Closes #10, #15)
The old message and anomaly system was inefficient and also didn't make much sense, so i've torn it apart and replaced it fully from the ground up

The new anomaly system requires you to define a few things:

MessageBoss, which handles communication from the framework to
both objects of the initial type (base input type of the manager, like CommandSender) and objects of the adapted type (objects desired by the framework user, like Player)

AnomalyCaptioner, which builds the base message objects for EACH anomaly

AnomalyPrinter, which adds messages to messages generated by the AnomalyCaptioner based on anomaly TYPE

and the AnomalyHandler, which delegates to the prior 3 in order to consume Anomalies and send them to the player.

The MessageBoss allows for you to use your own custom Message type, meaning that we can produce (MessageBoss, AnomalyCaptioner, AnomalyPrinter) implementations for existing user interfacing frameworks like Kyori-Adventure AND a homegrown implementation if users desire to not use existing frameworks.