Implement Block typing in anticipation of help scanning / syntax scanning
Not sure if this is the best way of going about things but i'm hoping it will prove an invaluable decision on the path to ~~destroying ACF~~ finishing work on Branch in order to phase out other command frameworks in the organization.

This commit also should phase out the use of blocks as a search argument - that's for strings: Instead, blocks are used as descriptive objects for help nodes and similar constructs to display information.

Currently working on implementing tests, continuing work on arguments, and soon implementing the multiadapter and help node.

Implementation of tests is extremely important to our future code quality and sustainability as it will allow for us to catch early bugs much faster as well as prototype and showcase working code. In the future, all pull requests for new features will require working tests with 60%+ test coverage. Additionally, in all additional projects i work on i will write tests alongside development as opposed to after completion since this is gonna be real tedious to do