Implement optionals
Added experimental support for optional arguments into the argument node system.

If an argument requires more than one string, it will only be substituted with an optional if **NO** strings are given. If an argument that requires 2 strings is given one and is optional, it will still throw a NoInputProvidedInternalAnomaly.